# Implement Rabin-Karp substring search.