--- a/Data/ByteString.hs
+++ b/Data/ByteString.hs
@@ -279,7 +279,7 @@ import GHC.Prim                 (Word#)
 import GHC.Base                 (build)
 import GHC.Word hiding (Word8)
 
-#if !(MIN_VERSION_BASE(4,7,0))
+#if !(MIN_VERSION_base(4,7,0))
 finiteBitSize :: Bits a => a -> Int
 finiteBitSize = bitSize
 #endif