--- a/bench/bench-bytestring.cabal
+++ b/bench/bench-bytestring.cabal
@@ -21,6 +21,8 @@ Tested-With:         GHC==7.0.3
 Build-Type:          Simple
 Cabal-Version:       >= 1.10
 
+data-files:          wiki-haskell.html
+
 source-repository head
   type:     darcs
   location: http://darcs.haskell.org/bytestring/
@@ -42,6 +44,7 @@ executable bench-bytestring-builder
                    -- we require bytestring due to benchmarking against
                    -- blaze-textual, which uses blaze-builder
                    , bytestring    >= 0.9
+                   , random
 
   -- cabal complains about ../ dirs. However, this is better than symlinks,
   -- which probably don't work on windows.