… bench/BenchAll.hs
… bench/wiki-haskell.html
… tests/Properties.hs