I have implemented Rabin-Karp substring search. This gives time complexity of O(mn) in the worst case (the same as before), but O(m + n) except for in very pathological cases.