@gregorycollins 2015-09-18
A Boyer-Moore would probably be better (or Boyer-Moore-Horspool, which I've also [implemented elsewhere](https://hackage.haskell.org/package/io-streams-1.3.2.0/docs/src/System-IO-Streams-Internal-Search.html#search)) I think, but this version is likely to be faster than what's there already so maybe we should take it.

Without accompanying benchmark results, however, I don't think there's any point in reviewing this -- so could you write an attoparsec benchmark or two comparing old vs new?

Also, please investigate the Travis CI failure.