@SeanRBurton 2015-09-23
Thanks for the feedback. Travis seems to be failing due to an infinite loop in the test suite, and I'm pretty sure it predates this pull request. I'll try and pin it down.