@SeanRBurton 2015-09-28
Here are some benchmark results:

New:
`
easy:

1:    142.6407 us 
4:    7.180527 ms 
16:   11.20857 ms 
64:   15.13670 ms 
128:  16.29927 ms 
1024: 11.35251 ms 

random:             
              
1:    779.7848 us 
4:    6.518944 ms 
16:   10.62446 ms                                          
64:   13.89249 ms                                          
128:  13.94245 ms                                          
1024: 12.68588 ms                                          

hard:                                                      
              
1:    42.34540 ms 
4:    8.812522 ms                                          
16:   13.88353 ms                                          
64:   10.71590 ms                                          
128:  12.14654 ms                                          
1024: 10.73882 ms                                          

pathological:                                              
              
4:    5.538884 ms                                          
16:   9.696482 ms                                          
64:   10.00006 ms                                          
128:  9.517451 ms 
1024: 9.473508 ms 

html:             
              
1:    759.9900 us 
4:    904.7559 us 
16:   1.560976 ms 
64:   1.562951 ms 
128:  1.566122 ms 
1024: 1.582036 ms 

Old:

easy:

1:    16.15ms     
4:    20.12ms  
16:   22.38ms  
64:   25.58ms  
128:  26.82ms   
1024: 28.63ms    

random:             

1:    17.18ms 
4:    22.00ms 
16:   25.58ms  
64:   27.08ms  
128:  27.68ms   
1024: 26.12ms    

hard:             
              
1:    27.27ms 
4:    24.80ms 
16:   27.34ms  
64:   31.08ms  
128:  31.61ms   
1024: 31.00ms    

pathological:             
              
1:    18.37ms 
4:    23.11ms 
16:   31.31ms  
64:   47.88ms  
128:  42.54ms                                     
1024: 156.8ms                                             

html:                                              
              
1:    2.989ms 
4:    3.585ms          
16:   3.933ms                                             
64:   4.950ms           
128:  5.126ms                     
1024: 4.810ms      

This change is clearly an improvement.