@gregorycollins 2015-09-28
Thanks for the benchmarks. Aside from the typo, this patch LGTM -- I will merge once the typo fix comes back unless I get objections. @tibbe @dcoutts