@tibbe 2015-09-28
LGTM. Thanks for the patch! Could you just please remove the whitespace changes or move them to separate commit?