@dcoutts 2015-09-30
LGTM. I presume our existing tests cover this ok, right?

If you're interested in comparing these algorithms, you might look at http://hackage.haskell.org/package/stringsearch which has a number of them. I think they also recommend their Boyer-Moore impl. But I'm happy to merge this as it's a strict improvement.