@dcoutts 2015-10-20
I'm happy to have it merged. If someone wants to fix up the conflicts that'd be perfect, otherwise I'll get to it eventually.